# app/api/swagger.yaml
openapi: 3.0.0
info:
  title: MealTime API
  version: 2.0.0
  description: |
    API completa para gerenciamento de alimentação de gatos
    
    **VERSÕES DISPONÍVEIS**:
    - **V2 (Recomendado)**: `/api/v2/*` - Autenticação JWT/Session, respostas padronizadas
    - **V1 (Deprecated)**: `/api/*` - Será removido em 2025-07-28
    
    **AUTENTICAÇÃO V2**:
    - Mobile: JWT via header `Authorization: Bearer <token>`
    - Web: Supabase Session via cookies
  contact:
    name: MealTime Support
    email: support@mealtime.app
servers:
  - url: http://localhost:3000/api/v2
    description: Servidor V2 (desenvolvimento) - RECOMENDADO
  - url: https://mealtime.app/api/v2
    description: Servidor V2 (produção) - RECOMENDADO
  - url: http://localhost:3000/api
    description: Servidor V1 (desenvolvimento) - DEPRECATED
  - url: https://mealtime.app/api
    description: Servidor V1 (produção) - DEPRECATED
...
paths:
...
  /v2/feedings:
    get:
      tags: [V2 - Feedings]
      summary: 🆕 Listar alimentações (V2)
      security:
        - bearerAuth: []
      parameters:
        - name: householdId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de alimentações
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2Response'
    
    post:
      tags: [V2 - Feedings]
      summary: 🆕 Registrar alimentação (V2)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [catId]
              properties:
                catId:
                  type: string
                  format: uuid
                amount:
                  type: number
                notes:
                  type: string
                meal_type:
                  type: string
                  enum: [manual, scheduled]
      responses:
        '201':
          description: Alimentação registrada
        '409':
          description: Alimentação duplicada (muito recente)
...
  /v2/feedings/stats:
    get:
      tags: [V2 - Feedings]
      summary: 🆕 Estatísticas de alimentação (V2)
      security:
        - bearerAuth: []
      parameters:
        - name: catId
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrar por gato específico (opcional)
        - name: days
          in: query
          schema:
            type: integer
            default: 7
            minimum: 1
            maximum: 90
      responses:
        '200':
          description: Estatísticas de alimentação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2Response'

  # ==================== V2 - WEIGHT LOGS (RECOMENDADO) ====================
...
  /cats/{catId}:
    get:
      tags: [Cats]
      summary: Obter gato por ID
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalhes do gato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
    
    put:
      tags: [Cats]
      summary: Atualizar gato
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatInput'
      responses:
        '200':
          description: Gato atualizado
    
    delete:
      tags: [Cats]
      summary: Deletar gato
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Gato deletado com sucesso
...
  /feedings:
    get:
      tags: [Feedings]
      summary: ⚠️ Listar alimentações (V1 - DEPRECATED)
      description: |
        **DEPRECATED**: Use `/v2/feedings` em vez disso.
        Será removido em 2025-07-28.
        
        Retorna histórico de alimentações
      deprecated: true
      parameters:
        - name: catId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Lista de alimentações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feeding'
    
    post:
      tags: [Feedings]
      summary: Registrar alimentação
      description: Cria um novo registro de alimentação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedingInput'
      responses:
        '201':
          description: Alimentação registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feeding'
...
  /feedings/stats:
    get:
      tags: [Feedings]
      summary: Estatísticas de alimentação
      description: Retorna estatísticas detalhadas sobre alimentações
      parameters:
        - name: catId
          in: query
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 7
      responses:
        '200':
          description: Estatísticas de alimentação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedingStats'

  # ==================== HOUSEHOLDS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtido via POST /api/auth/mobile
...
  schemas:
    V2Response:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indica se a requisição foi bem-sucedida
        data:
          type: object
          description: Dados retornados (quando success=true)
        count:
          type: integer
          description: Número de itens retornados (para listas)
        error:
          type: string
          description: Mensagem de erro (quando success=false)
        details:
          type: object
          description: Detalhes adicionais do erro (quando aplicável)
      example:
        success: true
        data:
          id: "uuid"
          name: "Exemplo"
        count: 1
    Cat:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        photoUrl:
          type: string
        birthDate:
          type: string
          format: date
        weight:
          type: number
        householdId:
          type: string
        ownerId:
          type: string
        portionSize:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CatInput:
      type: object
      required: [name, householdId]
      properties:
        name:
          type: string
        photoUrl:
          type: string
        birthDate:
          type: string
          format: date
        weight:
          type: number
        householdId:
          type: string
        portionSize:
          type: number
    Feeding:
      type: object
      properties:
        id:
          type: string
        catId:
          type: string
        userId:
          type: string
        timestamp:
          type: string
          format: date-time
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        portionSize:
          type: number
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    FeedingInput:
      type: object
      required: [catId, timestamp]
      properties:
        catId:
          type: string
        timestamp:
          type: string
          format: date-time
        mealType:
          type: string
        portionSize:
          type: number
        notes:
          type: string
    FeedingStats:
      type: object
      properties:
        period:
          type: object
          properties:
            days:
              type: integer
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
        totals:
          type: object
          properties:
            feedings:
              type: integer
            dailyAverage:
              type: number
        dailyStats:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              feedings:
                type: integer
        catStats:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              totalFeedings:
                type: integer
              dailyAverage:
                type: number
...
security:
  - bearerAuth: []